+++ bin/ncp/SYSTEM/nc-swapfile.sh
@@ -47,8 +47,7 @@ configure()
 
 install()
 {
-  apt-get update
-  DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends dphys-swapfile
+  apt_install dphys-swapfile
 }