+++ ncp.sh
@@ -34,7 +34,7 @@ install()
   # add the ncc shortcut
   cat > /usr/local/bin/ncc <<'EOF'
 #!/bin/bash
-sudo -u www-data php /var/www/nextcloud/occ "$@"
+sudo -E -u www-data php /var/www/nextcloud/occ "$@"
 EOF
   chmod +x /usr/local/bin/ncc
 
@@ -248,7 +248,8 @@ EOF
     echo nextcloudpi > /etc/hostname
 
     ## tag image
-    [[ -f /.docker-image ]] && local DOCKER_TAG="_docker"
+    is_docker && local DOCKER_TAG="_docker"
+    is_lxc && local DOCKER_TAG="_lxc"
     echo "NextCloudPi${DOCKER_TAG}_$( date  "+%m-%d-%y" )" > /usr/local/etc/ncp-baseimage
 
     ## SSH hardening